Reorganize Hamburger sample: shows now Default- and Creators Update style
